requirements.txt
bokeh==3.3.4
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
dask==2024.2.1
dask-jobqueue==0.8.5
distributed==2024.2.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
locket==1.0.0
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.2.1
partd==1.4.1
pillow==10.2.0
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sortedcontainers==2.4.0
sympy==1.12
tblib==3.0.0
threadpoolctl==3.5.0
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.4
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xyzservices==2023.10.1
zict==3.0.0
zipp==3.17.0
